Add the AugmentedMediaSource to the virtual camera sample. This source wraps an existing camera and fitlers the set of pins and MediaTypes it exposes out. It also supports a sample custom control that could be implemented to process pixels from frames passing through.

Also adds a SystemTray fulltrust process sidekick to the main UWP control app that launches when the main control app is closed. This serves 2 purpose: 1) offer an easier access to the app from system tray 2) have a background process listening for RPC call from the VirtualCamera to launch back the main control app when the virtual camera starts to stream. To accomodate this use case better, the main UWP control app is also request to register itself as a startup task on the system upon first launch and registrering a launch protocol for easier launch from the system tray process.
update tests, update names to use "Contoso" instead of "MS or Microsoft"

+++ Samples/VirtualCamera/VirtualCameraMediaSource/SimpleMediaStream.h
@@ -41,8 +41,8 @@ namespace winrt::WindowsSample::implementation
         DWORD Id() const { return m_dwStreamId; }
         MFSampleAllocatorUsage SampleAlloactorUsage() const { return m_allocatorUsage; }
 
-        void SetRGBMask(ULONG rgbMask) { winrt::slim_lock_guard lock(m_Lock);  m_rgbMask = rgbMask; }
-        ULONG GetRGBMask() { winrt::slim_lock_guard lock(m_Lock);  return m_rgbMask; }
+        void SetRGBMask(uint32_t rgbMask) { winrt::slim_lock_guard lock(m_Lock);  m_rgbMask = rgbMask; }
+        uint32_t GetRGBMask() { winrt::slim_lock_guard lock(m_Lock);  return m_rgbMask; }
 
     protected:
         HRESULT _CheckShutdownRequiresLock();
@@ -61,7 +61,7 @@ namespace winrt::WindowsSample::implementation
 
         bool m_bIsShutdown = false;
         MF_STREAM_STATE m_streamState = MF_STREAM_STATE_STOPPED;
-        ULONG m_rgbMask = KSPROPERTY_SIMPLEMEDIASOURCE_CUSTOMCONTROL_COLORMODE_BLUE;
+        uint32_t m_rgbMask = KSPROPERTY_SIMPLEMEDIASOURCE_CUSTOMCONTROL_COLORMODE_BLUE;
 
         DWORD m_dwStreamId;
         MFSampleAllocatorUsage m_allocatorUsage;
